+++ packages/hooks/src/data/QueryData.ts
@@ -467,10 +467,17 @@ export class QueryData<TData, TVariables> extends OperationData {
     ) => TData
   ) => this.currentObservable.query!.updateQuery(mapFn);
 
-  private obsStartPolling = (pollInterval: number) =>
-    this.currentObservable.query!.startPolling(pollInterval);
+  private obsStartPolling = (pollInterval: number) => {
+    this.currentObservable &&
+      this.currentObservable.query! &&
+      this.currentObservable.query!.startPolling(pollInterval);
+  };
 
-  private obsStopPolling = () => this.currentObservable.query!.stopPolling();
+  private obsStopPolling = () => {
+    this.currentObservable &&
+      this.currentObservable.query! &&
+      this.currentObservable.query!.stopPolling();
+  };
 
   private obsSubscribeToMore = <
     TSubscriptionData = TData,